dkgResultSigner and dkgResultSubmitter extracted to a separate file
Structs related to the result publication were extracted to
dkg_submit.go file. This way, it is easier to find components and tests
for them.
